[#62] Remove Assistant Chat Listener, merge some functionality
