+++ docker-compose.yml
@@ -7,7 +7,6 @@ x-environments: &rabbitmq_client_env
   RABBITMQ_EXCHANGE_USER_CHATS: ${RABBITMQ_EXCHANGE_USER_CHATS}
   RABBITMQ_QUEUE_USER_CHATS: ${RABBITMQ_QUEUE_USER_CHATS}
   RABBITMQ_QUEUE_USER_CHAT_REPLIES: ${RABBITMQ_QUEUE_USER_CHAT_REPLIES}
-  RABBITMQ_QUEUE_ASSISTANT_CHAT_LISTENER: ${RABBITMQ_QUEUE_ASSISTANT_CHAT_LISTENER}
   RABBITMQ_QUEUE_ASSISTANT_CHAT_REPLIES: ${RABBITMQ_QUEUE_ASSISTANT_CHAT_REPLIES}
   RABBITMQ_QUEUE_HISTORIES: ${RABBITMQ_QUEUE_HISTORIES}
